- Bug fixes: 1. fixed functional of products deleting from the cart (changed type of index in CartController.java) 2.fixed the relationship between OrderEntity.java & ProductEntity.java
- Changed:
1. Color of header.html

- Small fixes -

+++ src/main/java/org/example/FinalProject/controllers/CartController.java
@@ -40,7 +40,7 @@ public class CartController {
     @GetMapping("/cart/deleteFromCart/{index}")
     public String deleteFromCart(
             @ModelAttribute("cart") Cart cart,
-            @PathVariable("index") long index,
+            @PathVariable("index") int index,
             RedirectAttributes attributes) {
         cart.remove(index);
         attributes.addFlashAttribute("cart", cart);
